remove + implement: remove appointment remove and change to use read appointent repo and write appointment repo
